requirements.txt
dash==2.14.2
dash-bootstrap-components==1.6.0
dash-table==5.0.0
plotly==5.18.0
Flask==2.3.2
selenium==4.15.2
webdriver-manager==4.0.1
RPi.GPIO==0.7.1
apscheduler==3.10.4
gunicorn==21.2.0
sqlite==3.35.5
